+++ webclient/package.json
@@ -8,9 +8,6 @@
   "private": true,
   "proxy": "http://localhost:8080",
   "dependencies": {
-    "@testing-library/jest-dom": "^5.15.0",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-router-dom": "^5.1.2",
@@ -35,7 +32,6 @@
       ">0.2%",
       "not dead",
       "not op_mini all"
-      
     ],
     "development": [
       "last 1 chrome version",
