+++ webclient/public/index.html
@@ -1,34 +1,27 @@
 <html lang="en">
-  <head>
-    
-<script src="https://unpkg.com/peerjs@1.3.1/dist/peerjs.min.js" crossorigin="use-credentials"></script>
-  	<script type="text/javascript" src="./Room.js" crossorigin="use-credentials">
-document.write('<script src="http://' + (location.host || 'localhost').split(':')[0] +
-  ':35729/livereload.js?snipver=1"></' + 'script>')
-  </script>
-    <meta charset="utf-8" />
-    <link rel="icon" href="public/favicon.ico" crossorigin="use-credentials"/>
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta name="theme-color" content="#000000" />
-    <meta
-      name="description"
-      content="Web site created using create-react-app"
-    />
-    <link rel="apple-touch-icon" href="public/logo192.png" crossorigin="use-credentials"/>
-    <!--
+
+<head>
+  <meta charset="utf-8" />
+  <link rel="icon" href="public/favicon.ico" crossorigin="use-credentials" />
+  <meta name="viewport" content="width=device-width, initial-scale=1" />
+  <meta name="theme-color" content="#000000" />
+  <meta name="description" content="Web site created using create-react-app" />
+  <link rel="apple-touch-icon" href="public/logo192.png" crossorigin="use-credentials" />
+  <!--
       manifest.json provides metadata used when your web app is installed on a
       user's mobile device or desktop. See https://developers.google.com/web/fundamentals/web-app-manifest/
     -->
-    <link rel="manifest" href="public/manifest.json" crossorigin="anonymous"/>
-    <!--
+  <link rel="manifest" href="public/manifest.json" crossorigin="anonymous" />
+  <!--
       'public' is the %PUBLIC_URL%
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>React App</title>
-  </head>
-  <body>
-    <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div id="root"></div>
+  <title>React App</title>
+</head>
 
-  </body>
+<body>
+  <noscript>You need to enable JavaScript to run this app.</noscript>
+  <div id="root"></div>
+</body>
+
 </html>
