+++ webclient/read.md
@@ -1,0 +1,4 @@
+## How to bring this up (Web Client):
+* npm run build - will generate the build bundle to be run.
+* serve -s build 
+
