+++ webclient/src/App.js
@@ -1,24 +1,20 @@
 import React from "react";
 import { BrowserRouter, Route, Switch } from "react-router-dom";
-import CreateRoom from "CreateRoom";
-import Room from "Room";
-import express from "express";
-var router = express.Router();
+import CreateRoom from "./CreateRoom/index";
+import Room from "./Room/index";
 
 function App() {
   return (
-    <BrowserRouter>
-      <Switch>
-        <Route
-          path="/webclient/src/room/CreateRoom.js"
-          exact
-          component={CreateRoom}
-        />
-        <Route path="/webclient/src/room/room.js" exact component={Room} />
-      </Switch>
-    </BrowserRouter>
+      <BrowserRouter>   
+        <Switch>
+          <Route path="/create-room" component={CreateRoom} />
+          {
+          // TODO: room/${id} - Create the route that receives the uuid and mounts the video room for this single uuid room.
+          }
+          <Route path="/room" component={Room} />
+        </Switch>
+      </BrowserRouter>
   );
 }
 
-module.exports = router;
 export default App;
