+++ webclient/src/Room/styles.js
@@ -1,0 +1,17 @@
+import styled from "styled-components";
+
+const Container = styled.div`
+  padding: 20px;
+  display: flex;
+  height: 100vh;
+  width: 90%;
+  margin: auto;
+  flex-wrap: wrap;
+`;
+
+const StyledVideo = styled.video`
+  height: 40%;
+  width: 50%;
+`;
+
+export { Container, StyledVideo };
