+++ webclient/src/index.js
@@ -1,10 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import express from "express";
 import "./index.css";
 import App from "./App";
-import * as serviceWorker from "./serviceWorker";
-const router = express.Router();
 
 ReactDOM.render(
   <React.StrictMode>
@@ -12,7 +9,3 @@ ReactDOM.render(
   </React.StrictMode>,
   document.getElementById("root")
 );
-
-serviceWorker.unregister();
-
-module.exports = router;
